address issue #16149
Address the following issue:

https://github.com/files-community/Files/issues/16149

Avoid skipping detection of a drive named G that is not Google Drive when the
registry contains old keys for Google Drive.

Add "My Drive" to the Google Drive registry path in RemovableDrivesService and
check if it's valid.

Then DON'T skip the found "G" drive if the "My Drive" path is NOT valid.